0.2 release (cleanup and additional examples)
This is a new release.  In addition to cleaning up existing code,
it adds error checking (which identified a few bugs), and a number
of examples taken from my archived (unreleased) library.  This
includes an analysis of different saturation vapor pressure
formulations, the effect of the saturation vapor pressure formulation
on the potential temperatures, and differences in calculations of ths
lifting condensation level.