generic functions to invert state functions
By passing the function that is to be inverted it is possible to write a
generic inversion function, which we call invert_for_temperature and
invert_for_pressure respectively.  These work for all off the
thermodynamic funcions, i.e., theta_e, theta_l, etc.

I experimented with an invert_for_humidity, but this often did nto
converge so I did not introduce this functionality.