added a function to calculate (moist) static energy
in the atmospheric sciences there are many different moist static
energies.  They differ due to the choice of reference enthalpies for
water phases, which are connected by the phase change enthalpies, and
thus have a single degree of freedom.  This we give to the reference
enthalpy (enthalpy at T0) for vapor.  The choice of hv0 matters because
not all reference quantities can be zero for all phases of water, and so
the choice influences how the reference quantity weights different water
phases.

the function has been defined in a rather expressive way to bring out
this property and make it possible to compute one or the other of the
moist static energies by choosing a different reference state for the
vapor enthalpy.