restructure treatment of saturation vapor pressure
In this branch additional impelmentations of the saturation vapor
pressure were impelmented.  These different treatments are not meant to
be exhaustive, but the teten's murray formulations were introduced
because these are used by ICON.  The murphy-koop formulation is the
standard for super-cooled liquid water.  The examples ipynb script was
augmented to present the new functionality.