revised analytic saturation vapor pressures
Attempted to clean up the presentation of the Teten's and Romps formulae
for saturation vapor pressuer, incorporating improved fitting constants
for the former.  These were derived by fitting to Wagner and Pruss for
the temperature range 270 K to 310 K and to Wagner et al for ice over
the range 230K to 260 K.  The fitting is not particular sensitive and
fitting with noice leads to slight changes in the fits, but the new
constants are robustly an improvement over the old ones, which had
inconsistent triple points vapor pressures over ice and liquid. I also
double checked the 'Romp's' formulae, and refrained from calling these
Romps as they are just straightforward integrations of the
Clausius-Clapeyron equation under the Rankine-Kirchoff assumptions.