revised generic fuctions as closures
To define generic functions we use the closure construct which
introduces a way to make a function with the given parameters so that
when it is called it is executed with those parameters without having to
pass the parameters.  This allows the function calls for different
quantities to share a common interface, and more allows for a cleaner
separation between a function and its parameters, the latter being used
to construct the function.